First commit of my solution. I have initialized npm so I am using npm to manage my dependencies.
So far in my solution I:

1. Load the contents of the output file if there are any.
2. Read the contents of thew CSV file one line at a time.
3. Skip any entries in the CSV file that already have results
   in the output file based on store number.

The problem requires that I call an HTTP GET API endpoint to get
geolocation information for the given store number based on the store's
address.

I have started to implement that. I have a Node.JS function named
geolocateStoreAddress which uses the axios package module to make an
HTTP GET request to the endpoint with the required parameters.

The axios call has a .then which is called if the HTTP GET request
succeeds and a .catch that gets called if the HTTP GET request fails.

I have started a getHttpResponseHandler which is a JavaScript function
that returns a callback (JavaScript function for processing the
HTTP response from the HTTP GET request). I use that approach because
I need the storeNumber and the address as a part of handling the
response though there may be an easier/more straight forward way to handle
this issue.