I have previously implemented a main function in my Node.JS solution. Node.JS does not require a main function definition. I have done this for clarity and for scoping.
In the previous commit if the output JSON file already exists, it loads
the contents of that output JSON file to avoid recomputing the
geolocation data for stores for which it has already done the work.

In this commit I am creating a JSON string from the data that I
loaded from the output JSON file. I generate that string by calling
JSON.stringify on the JSON object that I parsed from the output file.
I then throw an exception if that string does not end in right curly brace.

The reason I am doing this should become more clear in a few commits.