I did a small refactor: in my response handler I need to get the response payload and if it is empty, then I want to report a warning via console.warn and I want to skip that given store from my output JSON file.
To that end I implemented extractResponsePayload which takes
in the response. If the response has no payload, null is returned.
Otherwise the response payload is returned in the form of a
JavaScript object that contains the geolocation data for the
store.

The geolocation data includes a type property that is the
type of the query performed, for the given input CSV file we
have it always has the value 'address'. I think that is
confusing so I dropped the type property from the geolocation
data.

My extractResponsePayload uses lodash.isempty to test
the JavaScript object to see if it is empty. So I also
took a dependency on lodash.isempty.