On a previous commit I had started to remove the code that writes the output JSON data into the output JSON file at the start of execution.
Previously, I was loading in the output JSON data from the
output JSON file and then writing immediately back into the
file.

That does not make sense.

So I stripped out that code. However, unfortunately I missed
some of it. In this commit I got the rest of the code out.